+++ NEWS.md
@@ -1,5 +1,8 @@
 # paradox 1.0.1-9000
 
+* `ParamSetCollection$flatten()` now detaches `$extra_trafo` completely from original ParamSetCollection.
+* Option to postfix, instead of prefix, in `ParamSetCollection`, `c()`/`ps_union()`, and `ps_replicate()`.
+
 # paradox 1.0.1
 
 * Performance improvements.